Added shield and double shot pickups

+++ Never-tell-me-the-odds/Assets/Scripts/Components/Settings/BulletSettingsComponent.cs
@@ -7,6 +7,6 @@ public struct BulletSettingsComponent : IComponentData
 {
     //bullets
     public float MaxBulletAge;
-    public float MaxBulletCount;
+    public int MaxBulletCount;
     public float BulletBaseSpeed;
 }
